requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
idna==3.7
Naked==0.1.32
phonenumbers==8.13.35
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
PyJWE==1.0.0
PyJWT==2.1.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
shellescape==3.8.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
